+++ contracts/base/ClearingHouseCallee.sol
@@ -5,16 +5,13 @@ pragma abicoder v2;
 import { IUniswapV3Pool } from "@uniswap/v3-core/contracts/interfaces/IUniswapV3Pool.sol";
 import { AddressUpgradeable } from "@openzeppelin/contracts-upgradeable/utils/AddressUpgradeable.sol";
 import { SafeOwnable } from "./SafeOwnable.sol";
-import { MarketRegistry } from "../MarketRegistry.sol";
 
-// TODO split to chCallee and UniswapCallbackBridge
 abstract contract ClearingHouseCallee is SafeOwnable {
     using AddressUpgradeable for address;
 
     //
     // STATE
     //
-    address public marketRegistry;
     address public clearingHouse;
 
     // __gap is reserved storage
@@ -34,22 +31,11 @@ abstract contract ClearingHouseCallee is SafeOwnable {
         _;
     }
 
-    modifier checkCallback() {
-        address pool = _msgSender();
-        address baseToken = IUniswapV3Pool(pool).token0();
-        require(pool == MarketRegistry(marketRegistry).getPool(baseToken), "EX_FCV");
-        _;
-    }
-
     //
     // CONSTRUCTOR
     //
-    function __ClearingHouseCallee_init(address marketRegistryArg) internal initializer {
+    function __ClearingHouseCallee_init() internal initializer {
         __SafeOwnable_init();
-
-        // MarketRegistry is not contract
-        require(marketRegistryArg.isContract(), "CHD_MRNC");
-        marketRegistry = marketRegistryArg;
     }
 
     function setClearingHouse(address clearingHouseArg) external onlyOwner {